ccn-lite: shell: do not count white space after single word
If a single word is added to the content cache via `ccnl_cont /TEST
hello`, then the resulting wireshark dump of a successful `ccnl_int
/TEST` will contain a TLV of len(hello) plus 1 (whit space). By
incrementing `buf_ptr` only for loop iterations `> 2` fixes this. As a
bonus, I removed the superfluous `memset` to white space, as the buffer
is correctly terminated with a '\0' character.