examples: only whitelist a few essential boards (dirty hack)
This is a dirty hack proposed by @kaspar030 offline (with slight
adaptations) to allow Murdock to build properly tonight. @kaspar030
revert if the issue is fixed, @miri64 close this one if it isn't merged
tonight.

Right now Murdock fails randomly for some boards, because it doesn't has
the `xxd` tool installed.