testrunner: make traceback extraction downwards-compatible
The object-like access to frame information in the traceback was only
introduced in Python 3.5. Before that version it was a 4-tuple [[1]].

The indexed way to access the frame seems to be upwards-compatible for
newer versions (tested with python 3.5, maybe some of the Arch crew can
test with even newer versions), so I used that one.

[1]: https://docs.python.org/3.4/library/traceback.html#traceback.extract_tb