gnrc_pktbuf_cmd: module is not a pktbuf implementation
When I introduced this pseudo-module in #9391 I did not consider, that
the way I named it, it is assumed to be an implementation of the
`gnrc_pktbuf` interface. However, it is not (it just provides the shell
command).

This PR excludes `gnrc_pktbuf_cmd` from the selection algorithm, so
that if it is the only `gnrc_pktbuf_%` module, still
`gnrc_pktbuf_static` gets selected.