tests/periph_uart: adapt main.c for automated testing
This is an effort to automate UART testing using the same framework as
was already introduced for i2c. The framework relies on shell commands
to return invocation status and data in a special way:

1. Invocation status: Success or Error
2. Data inside the []

Hence this patch adds invocation status to the UART init command and
invocation status and data to the printer.

Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>