makefiles/scan-build: Fix 'scan-build-analyze' target
This correctly defines a `scan-build-analyze` target that does not display the
result webpage.

`scan-build-view` has now been moved to a private target as should not
be used directly.

The handling of displaying the page on the host system and implementing
'scan-build-analyze' are now explicitely done with separate targets and
not double implemented target when in docker or on the host that were executed
twice with different implementations.