boards/nz32-sc151: do not blindly set TERMFLAGS
The board should not globally set `TERMFLAGS` without a RIOT_TERMINAL
condition.

Also the board does not export uart through usb and thus needs an uart
converter connected. Other boards in the same situation do not need
to set `TERMPROG` without a valid baudrate. So remove the definition
anyway.

This also removes the export of TERM* variables and should make it work
with the other RIOT_TERMINAL.