boards/msba2,avsextreme: Removed defunct code
The MSB-A2 and the Avsextrem board once had support for using the USB
interface of the LPC2387. The code setting up the USB clock has been unused
for ages and is now defunct, as the required value for `USBCLKDivValue` is no
longer present in RIOT's code base. This commits removes the defunct and unused
code.