.murdock: blacklist the -gz (debug compression) compiler flag.
-gz enables DWARF section compression, though it is a no-op where debug
is not enable, e.g. in murdock.

Due to a combination of factors, including a bug in CCACHE, this flag
makes CCACHE ineffective and slows down murdock.

The alternative, of filtering the flags in our buid system (for example,
by testing RIOT_CI_BUILD or CCACHE) are too much of a shotgun approach
as the problem is quite specific to our CI setup.

Read the comment in the file for a more in-depth description.