tests/nimble_l2cap_server: change order in _on_data
Printing the 'Received ...' string takes a short while and it is possible
that data is received while the string is printed. It seems however that
NimBLE does not like to be without a mbuf ready for taking data while
receiving something, as this seems to lead to a memory leak somehow. Now
changing the order of actions inside the _on_data() function fixes this.