makefiles: clean up default-channel.inc.mk
Clean up `default-channel.inc.mk` to only set the custom CFLAGS
if they are really needed.
This should help avoid cluttering `riotbuild.h`.

The default PAN ID configuration should be just as easy to set
as the default channel, so also add an option for that while we
are at it.