Revert "drivers/hd44780: add short delay after each character"
This reverts commit 68352fde3c50399b5b2815b3235fab3a516e88b2.

Adding an unexplained delay seemed wrong in the first place, but it fixed
the display on the MCB2388.

Turns out the display was erroneously operating in 8-bit mode due to the
uninitialized flag register.
Why the delay helped here I don't know.

But with #12634 fixing this, this hack is not needed anymore.