tests/sys_arduino: fix test synchronization issues
The test currently doesn't wait for synchronization of the serial
terminal the second time it is setup with `Serial.begin`. This results
in the test commands being sent before the terminal is ready which
causes the test to fail. To fix this add an expect_exact on "Hello
arduino". (NOTE: this is different form `interactive_sync` which
synchronizes the start of the test, here we want `uart` to properly
re-init)