sc_gnrc_netif: fix whitespaces in output
Originally, the options and flags in the `netif` shell output were
separated by two spaces. For later added flags this is not the case,
making the parsing of those flags and options hard to impossible.

This change adds those missing spaces + comments so it might not happen
again in the future.