tools/suit: Add optional component ID argument
This extends the component argument of the gen_manifest.py script with
an optional component name. As component names in the SUIT manifest
consist of an array of bytestrings, the commponent name in the argument
is also split. Each part of the name ends up as a separate part of the
component name in the manifest. Component names are split by colons to
match the separator used in the rest of the argument.