tests/pkg_lvgl: fix sysmon task creation
Fixes a bug on the `pkg_lvgl` test when running on the esp32. When `sysmon_task`
was created it ran too early before any widget was created causing
Store/LoadProhibited exceptions on the esp32, I assume it's noticeable on the
esp32 due to the code being fetched from external flash memory adding a delay
higher than `REFR_TIME`.

Signed-off-by: Jean Pierre Dudey <me@jeandudey.tech>