pkg/hacl: switch upstream to backup copy
The https://github.com/mitls/hacl-c repo was deleted.
Getting the new upstream (https://github.com/project-everest/hacl-star)
to work nicely as a package requires some non-trivial work and in the
meantime all CI jobs will fail.

As a short-term solution, switch the upstream to a backup version of
the original repository.